Speed up git operations by multi-threads
Gilt clone and extract/overlay git resources in sequence, it's
slow when there's a lot gilt resources defined in config. In this
commit, it uses multi-threading to speed up these operations, in
order to make the thread operations safe, it uses git environment
variables instead of chdir into the repo clone/branched.