Determine branch in a smarter way
The git module is determine if a version is a branch by regrex
check if it's a sha, that makes it confuse on tag vs branch.
This commit determines branch by check return code of
`git show-ref --verify --quiet refs/heads/<version>`, this will
always work because gilt always fetch and checkout, there must
have heads ref to a branch.