Only fetch when branch/tag/commit not on local repo
The fetch operation is a slow operation, if a commit/tag/branch
has been on local repo, we can skip the fetch operation.
For branch, if it's a known branch, the following pull will also
fetch from remote.
For tag/commit, if it's a know local commit/tag, the fetch is
not required as it has been downloaded to local.