Fix tox failure due to use of relative path with --cov
Move py.test params to setup.cfg in order to make them available
to all pytest executions, regardless if they are run inside or ourside
tox.

This fixes problem where tox was failing to create coverage because
{toxinidir} was expanded to a relative path which is not supported
by pytest-cov.

By using the path mentioned in setup.cfg we avoid this issue as
in this case pytest-cov is able to generate the files.